refactor: error handling comments part 1
