+++ pop-api/examples/fungibles/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 
 [dependencies]
 ink = { version = "5.0.0", default-features = false }
-pop-api = { path = "../../../pop-api", default-features = false }
+pop-api = { path = "../../../pop-api", default-features = false, features = ["assets"] }
 scale = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive"] }
 scale-info = { version = "2.6", default-features = false, features = ["derive"], optional = true }
