+++ pop-api/src/lib.rs
@@ -1,12 +1,19 @@
 #![cfg_attr(not(feature = "std"), no_std, no_main)]
 
-use crate::error::PopApiError;
 use ink::{prelude::vec::Vec, ChainExtensionInstance};
-use primitives::{cross_chain::*, storage_keys::*, AccountId as AccountId32};
-use scale::Encode;
 pub use sp_runtime::{BoundedVec, MultiAddress, MultiSignature};
-use v0::RuntimeCall;
-pub use v0::{assets, balances, cross_chain, nfts, relay_chain_block_number, state};
+
+use crate::error::{PopApiError, StatusCode};
+use primitives::{storage_keys::*, AccountId as AccountId32};
+#[cfg(feature = "assets")]
+pub use v0::assets;
+#[cfg(feature = "balances")]
+pub use v0::balances;
+#[cfg(feature = "cross-chain")]
+pub use v0::cross_chain;
+#[cfg(feature = "nfts")]
+pub use v0::nfts;
+use v0::{state, RuntimeCall};
 
 pub mod error;
 pub mod primitives;
@@ -15,11 +22,7 @@ pub mod v0;
 type AccountId = AccountId32;
 // TODO: do the same as the AccountId above and check expanded macro code.
 type Balance = <Environment as ink::env::Environment>::Balance;
-type BlockNumber = <Environment as ink::env::Environment>::BlockNumber;
-type StringLimit = u32;
-type MaxTips = u32;
-
-pub type Result<T> = core::result::Result<T, PopApiError>;
+pub type Result<T> = core::result::Result<T, StatusCode>;
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 #[cfg_attr(feature = "std", derive(scale_info::TypeInfo))]
@@ -40,7 +43,7 @@ impl ink::env::Environment for Environment {
 
 #[ink::chain_extension(extension = 909)]
 pub trait PopApi {
-	type ErrorCode = PopApiError;
+	type ErrorCode = StatusCode;
 
 	#[ink(function = 0)]
 	#[allow(private_interfaces)]
@@ -50,6 +53,7 @@ pub trait PopApi {
 	#[allow(private_interfaces)]
 	fn read_state(key: RuntimeStateKeys) -> Result<Vec<u8>>;
 
+	#[cfg(feature = "cross-chain")]
 	#[ink(function = 2)]
 	#[allow(private_interfaces)]
 	fn send_xcm(xcm: CrossChainMessage) -> Result<()>;
@@ -67,6 +71,7 @@ fn read_state(key: RuntimeStateKeys) -> Result<Vec<u8>> {
 	.read_state(key)
 }
 
+#[cfg(feature = "cross-chain")]
 fn send_xcm(xcm: CrossChainMessage) -> Result<()> {
 	<<Environment as ink::env::Environment>::ChainExtension as ChainExtensionInstance>::instantiate(
 	)
