--- pop-api/src/v0/assets/pallets/mod.rs
@@ -1,1 +1,0 @@
-pub(crate) mod assets;
