--- pop-api/src/v0/assets/use_cases/mod.rs
@@ -1,1 +1,0 @@
-pub mod fungibles;
