+++ runtime/devnet/Cargo.toml
@@ -91,7 +91,7 @@ parachain-info.workspace = true
 env_logger = "0.11.2"
 hex = "0.4.3"
 enumflags2 = "0.7.9"
-pop-api = { path = "../../pop-api", default-features = false }
+pop-api = { path = "../../pop-api", defeult-features = false, features = ["assets"] }
 
 [features]
 default = ["std"]
